Map distance computation prototype.
The map geometry generation script now puts its output into the
generated data directory. Instead, a new script combines that geometry
data with the amenity data to generate the final map data, which
is then stored into the public data directory.

This public map data has a set of distances per amenity type for each
feature of the map. An amenity type is only present if at least one
amenity of that type exists within a 2.5 km radius of the feature's
center. This distance is stored in metres.

For now, the distance calculation is the euclidean distance. Doing it
for walking distances with OpenRouteService is out of scope here.

Also did some quick wip display of map colour, only for the food store
amenity type, using some very dumb distance -> score heuristic.